[sui-move] Unbreak build
